+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Enh #261: Enhance `RadioList::addRadioWrapClass()` method for cleaner class addition (@vjik)
 - Enh #263: Explicitly import classes and constants in "use" section (@mspirkov)
 - Chg #243: Deprecate static constructors such as `TagName::tag()` in favor of `new TagName()` (@razvbir)
+- Chg #221: Throw `LogicException` in `Tag::id()` when id is empty string (@razvbir)
 
 ## 3.12.0 December 13, 2025